requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sentence-transformers
psycopg2-binary
sqlalchemy
pgvector
pydantic